feat: Deve retornar erro se o item não existir

+++ src/services/cart-items/remove-item-to-cart.service.ts
@@ -1,4 +1,5 @@
 import CartItemsRepository from '../../repositories/cart-items.repository';
+import { MissingParamError } from '../../utils/errors/missingParamError';
 
 class RemoveItemToCartService implements RemoveItemToCartService {
   private readonly cartItemsRepository: CartItemsRepository;
@@ -8,6 +9,9 @@ class RemoveItemToCartService implements RemoveItemToCartService {
   }
 
   async removeItem(id: string): Promise<boolean> {
+    const existItem = await this.cartItemsRepository.getItemByID(id);
+    if (!existItem) throw new MissingParamError('Item inexistente', 401);
+
     return await this.cartItemsRepository.removeItem(id);
   }
 }
